backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "emails (>=0.6,<0.7)",
    "pytz (>=2025.2,<2026.0)",
    "attrs (>=25.3.0,<26.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "cryptography (>=45.0.3,<46.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
coverage = "^7.7.1"
ruff = "^0.11.2"
pytest = "^8.3.5"
